+++ Cheweishi/app/src/main/java/com/mining/app/zxing/view/ViewfinderView.java
@@ -199,7 +199,7 @@ public final class ViewfinderView extends View {
 			paint.setTextSize(TEXT_SIZE * density);
 			paint.setAlpha(0x40);
 			paint.setTypeface(Typeface.create("System", Typeface.NORMAL));
-			canvas.drawText(getResources().getString(R.string.scan_text), frame.left, (float) (frame.bottom + (float)TEXT_PADDING_TOP *density), paint);
+			canvas.drawText(getResources().getString(R.string.scan_text), frame.left, frame.bottom + (float)TEXT_PADDING_TOP *density, paint);
